core-9215 Refactor code.
