+++ de-lib/src/main/java/org/iplantc/de/desktop/client/views/ReactDesktop.java
@@ -25,6 +25,7 @@ public class ReactDesktop {
         DesktopView.Presenter presenter;
         MessageServiceFacade messageServiceFacade;
         Splittable[] windowConfigList;
+        String desktopContainerId;
     }
 
 }
