+++ react-components/src/desktop/view/Notifications.js
@@ -37,10 +37,6 @@ class Notifications extends Component {
         this.props.notificationClicked(event.currentTarget.id);
     }
 
-    updateApplicationTitle(count) {
-        document.title = getMessage("cyverseDe", {values: {count: count}})
-    }
-
     render() {
         const {
             anchorEl,
@@ -48,16 +44,17 @@ class Notifications extends Component {
         const notifications = this.props.notifications;
         const unSeenCount = this.props.unSeenCount;
         const messages = (notifications && notifications.messages && notifications.messages.length > 0) ? notifications.messages : [];
-        this.updateApplicationTitle(unSeenCount);
         return (
                 <span>
                       <img className={css(styles.menuIcon)}
                            src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAGQAAABkCAMAAABHPGVmAAAACXBIWXMAAC4jAAAuIwF4pT92AAAAGXRFWHRTb2Z0d2FyZQBBZG9iZSBJbWFnZVJlYWR5ccllPAAAAGBQTFRF////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////BmKR8wAAAB90Uk5TAAYJEhwiJ1VYWWR5e4CJjpWWp7HJyuXm7e7w8ff6/lycuGgAAAD0SURBVHgB7c1LsoJADAXQx1/5NNqoaZXm7n+Xz7IYMNCCFJOkzNnA+fvMGGOSxDl6cS5JdCfOYeacJauIMCPSnWBBb1IUTYOFpikKdUlZDsOAD4ahLEs9SV3HiG9ijHWtI2lbrGpb+UlVTRNWTVNVyU6y7H7HJiFkmeSk77FZ38tN0vT5xGaPR5pKTY5HsBwOUpPzGSyn008n1ytYLhepCRFYbjepyTiCZRylJmCz5M0SSyyxxBJLLOk6sHWdtIRf8Jv9Cb/gNwKS7wW/EZBgtx2JJdoZIiwQ6U28x4L3epM8DwGzEPLcktXGe3rx/l1oS4wx/4jHcRyd/1M5AAAAAElFTkSuQmCC"
                            alt="Notifications" onClick={this.handleNotificationsClick}></img>
+                    {unSeenCount !== "0" &&
                     <span id='notifyCount'
                           className={css(styles.unSeenCount)}>
-                        {(unSeenCount === "0") ? "" : unSeenCount}
+                        {unSeenCount}
                     </span>
+                    }
                     <Menu id={build(ids.DESKTOP, ids.NOTIFICATIONS_MENU)}
                           anchorEl={anchorEl}
                           open={Boolean(anchorEl)}
